requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
supabase==2.10.0
uvicorn==0.34.0